fastapi-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
qdrant-client==1.9.0
openai>=1.0.0
python-dotenv==1.0.1 